Add a few more commands to get A11.fsb to work
I'm starting to feel the pain of not properly tracking the references
as any kind of graph. I *can* technically keep implementing operators
and data types, but in order to actually understand the control flow, I
need to start making a proper decompiler that matches against patterns
of label-less gotos.